requirements.txt
streamlit
PyPDF2
python-docx
groq
scikit-learn
nltk
python-dotenv
